+++ src/controllers/userController.ts
@@ -1,22 +1,7 @@
 // Importing the necessary interfaces from the express module
-import { Request, Response, NextFunction } from "express";
+import { Request, Response } from "express";
 
 // Users controllers
-exports.checkId = (
-	req: Request,
-	res: Response,
-	next: NextFunction,
-	value: string | number
-) => {
-	console.log(`ID received: ${value}`);
-	console.log(`Check if ID is valid or not.`);
-	next();
-};
-
-exports.checkBody = (req: Request, res: Response, next: NextFunction) => {
-	console.log(`Check for required parameters in the request body.`);
-	next();
-};
 
 exports.getUsers = (req: Request, res: Response) => {
 	res.status(200).json({